refactor: separate filtering logic
